Refactoring des Writers
Issue #1
