+++ src/ArgeHeiwako.IO/ArgeHeiwako.IO.csproj
@@ -40,7 +40,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="DataFileWriter.cs" />
+    <Compile Include="OrdnungsbegriffeWriter.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
